Created the createFormObjectInnerHTML.js, this one is focused on server side, writes the html form as a string into a single property of the object.  When the form is built that property can be used by a variable or passed.  It can be applied as innerHTML to an existing html element.
updated some on creatForm and edited the createFormObject name to createFormElement as this better describes the objects purpose.  It will create a form using the createElement method.  Then will organize these elements into the proper order using parentElementId property of the form element object.  This is client side code as the DOM is not accessible on the server side.

added script to html index page in order to be able to swap between js files for testing purposes.